Split samples for zonal location
